+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+# 0.2.0
+* Add optional `expiration` parameter in the `Cache.set` method. This only affects the `TlruCache` and when specified, this takes precedence over the common expiry duration.
+
 # 0.1.2
 
-* Evict expired entry and return null. #1
+* Evict expired entry and return null. [#1](https://github.com/happy-san/dart-cacher/issues/1)
 
 # 0.1.1+1